+++ src/Candoumbe.Pipelines/Components/IHaveCoverage.cs
@@ -5,16 +5,16 @@ namespace Candoumbe.Pipelines.Components;
 /// <summary>
 /// Defines directories used for reporting code coverage
 /// </summary>
-public interface IHaveCoverage : IHaveOutputDirectory
+public interface IHaveCoverage : IHaveReport
 {
     /// <summary>
     /// Directory where to report code coverage
     /// </summary>
-    public AbsolutePath CoverageReportDirectory => OutputDirectory / "coverage-report";
+    public AbsolutePath CoverageReportDirectory => ReportDirectory / "coverage-report";
 
     /// <summary>
-    /// Directory where to publish converage history report
+    /// Directory where to publish code coverage history report
     /// </summary>
-    public AbsolutePath CoverageReportHistoryDirectory => OutputDirectory / "coverage-history";
+    public AbsolutePath CoverageReportHistoryDirectory => ReportDirectory / "coverage-history";
 
 }